fix(analysis): stop widening whole-number attributes to doubles
GSON's default number policy deserializes every JSON number as a Double, so any
read-then-write round trip rewrote an integer attribute `1` as `1.0`. A cc.json
that passed through a filter therefore no longer matched the file it came from,
and the drift compounded across a merge chain.

Both readers switch to ToNumberPolicy.LONG_OR_DOUBLE, which keeps integral
values as Long. CoberturaStrategy's casts are hardened to (as Number).toDouble()
because they can now see either type. The SVN golden file and the Tokei and
DependaCharta expectations are updated to the values that are actually produced.

+++ analysis/model/src/test/kotlin/de/maibornwolff/codecharta/serialization/NodeJsonDeserializerTest.kt
@@ -110,7 +110,7 @@ class NodeJsonDeserializerTest {
         val node = deserializer.deserialize(obj, nodeClass, null)
 
         assertThat(node.link).isEqualTo(url)
-        assertThat(node.attributes[attributeName] as Double).isEqualTo(attributeValue.toDouble())
+        assertThat(node.attributes[attributeName] as Long).isEqualTo(attributeValue.toLong())
         assertThat(node.children.size).isEqualTo(1)
     }
 }
